+++ Example/AETransitionExample/SourceViewController.swift
@@ -28,28 +28,32 @@ final class SourceViewController: UIViewController {
     var randomPresenting: AnimatedTransition {
         let index = Int.random(min: 0, max: presentingTransitions.count - 1)
         let random = presentingTransitions[index]
-        print(random.debugDescription)
+        print(random.debugDescription ?? "")
         return random
     }
 
     var randomDismissing: AnimatedTransition {
         let index = Int.random(min: 0, max: dismissingTransitions.count - 1)
         let random = dismissingTransitions[index]
-        print(random.debugDescription)
+        print(random.debugDescription ?? "")
         return random
     }
 
     var presentingTransitions: [AnimatedTransition] {
         return [
-            Transition.FadeIn(options: .random()),
-            Transition.MoveIn(from: .random(), options: .random())
+            Transition.FadeIn(crossfade: Bool.random(), options: .random()),
+            Transition.TranslateIn(from: .random(), push: Bool.random(), options: .random()),
+            Transition.RotateIn(angle: .pi, options: .random()),
+            Transition.ScaleIn(x: CGFloat.random(min: 0, max: 2), y: CGFloat.random(min: 0, max: 2), options: .random())
         ]
     }
 
     var dismissingTransitions: [AnimatedTransition] {
         return [
-            Transition.FadeOut(options: .random()),
-            Transition.MoveOut(to: .random(), options: .random())
+            Transition.FadeOut(crossfade: Bool.random(), options: .random()),
+            Transition.TranslateOut(to: .random(), push: Bool.random(), options: .random()),
+            Transition.RotateOut(angle: .pi, fadeOut: Bool.random(), options: .random()),
+            Transition.ScaleOut(x: CGFloat.random(min: 0, max: 2), y: CGFloat.random(min: 0, max: 2), fadeOut: Bool.random(), options: .random())
         ]
     }